pyproject.toml
[tool.poetry]
name = "siena_mls"
version = "18"
description = "Multimedia Environment for Students"
authors = ["Robin Flatland <flatland@siena.edu>","Ninad Chaudhari <nchaudhari@siena.edu>"]
license = "aGPLv3"
readme = "README.md"

homepage = "https://csis110.com/mls"
repository = "https://github.com/Introduction-To-CS-Siena-College/Siena-MLS-py"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^10.1.0"
scipy = "^1.11.3"
numpy = "^1.26.1"
python-resize-image = "^1.1.20"
pi-heif = "^0.13.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
coverage= "^7.3.2"
pytest-cov = "^4.1.0"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/Introduction-To-CS-Siena-College/Siena-MLS-py/issues"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-s --cov siena_mls --cov-report term --cov-report html"
testpaths = [
    "tests"
]
pythonpath = [
 "tests"
]
